+++ src/Resizer/ResizerInterface.php
@@ -20,18 +20,11 @@ use Sonata\MediaBundle\Model\MediaInterface;
 interface ResizerInterface
 {
     /**
-     * @param MediaInterface $media
-     * @param File           $in
-     * @param File           $out
-     * @param string         $format
-     * @param array          $settings
+     * @param string $format
      */
     public function resize(MediaInterface $media, File $in, File $out, $format, array $settings);
 
     /**
-     * @param MediaInterface $media
-     * @param array          $settings
-     *
      * @return Box
      */
     public function getBox(MediaInterface $media, array $settings);