+++ src/Security/PublicDownloadStrategy.php
@@ -24,9 +24,6 @@ class PublicDownloadStrategy implements DownloadStrategyInterface
      */
     protected $translator;
 
-    /**
-     * @param TranslatorInterface $translator
-     */
     public function __construct(TranslatorInterface $translator)
     {
         $this->translator = $translator;