+++ tests/CDN/PantherPortalTest.php
@@ -27,7 +27,7 @@ class PantherPortalTest extends TestCase
             '',
             false
         );
-        $client->expects($this->exactly(3))->method('flush')->will($this->returnValue('Flush successfully submitted.'));
+        $client->expects($this->exactly(3))->method('flush')->willReturn('Flush successfully submitted.');
 
         $panther = new PantherPortal('/foo', 'login', 'pass', 42);
         $panther->setClient($client);
@@ -53,7 +53,7 @@ class PantherPortalTest extends TestCase
             '',
             false
         );
-        $client->expects($this->exactly(1))->method('flush')->will($this->returnValue('Failed!!'));
+        $client->expects($this->exactly(1))->method('flush')->willReturn('Failed!!');
 
         $panther = new PantherPortal('/foo', 'login', 'pass', 42);
         $panther->setClient($client);