+++ tests/Form/Type/AbstractTypeTest.php
@@ -52,11 +52,11 @@ abstract class AbstractTypeTest extends TypeTestCase
         $this->formBuilder
             ->expects($this->any())
             ->method('add')
-            ->will($this->returnCallback(function ($name, $type = null): void {
+            ->willReturnCallback(function ($name, $type = null): void {
                 if (null !== $type) {
                     $this->assertTrue(class_exists($type), sprintf('Unable to ensure %s is a FQCN', $type));
                 }
-            }));
+            });
 
         $this->formType = $this->getTestedInstance();
     }