+++ tests/Form/Type/MediaTypeTest.php
@@ -47,14 +47,14 @@ class MediaTypeTest extends AbstractTypeTest
 
     public function testMissingFormOptions(): void
     {
-        $this->mediaPool->expects($this->any())->method('getProviderList')->will($this->returnValue([
+        $this->mediaPool->expects($this->any())->method('getProviderList')->willReturn([
             'provider_a' => 'provider_a',
             'provider_b' => 'provider_b',
-        ]));
-        $this->mediaPool->expects($this->any())->method('getContexts')->will($this->returnValue([
+        ]);
+        $this->mediaPool->expects($this->any())->method('getContexts')->willReturn([
             'video' => [],
             'pic' => [],
-        ]));
+        ]);
 
         $this->expectException(MissingOptionsException::class);
         $this->expectExceptionMessage(
@@ -66,14 +66,14 @@ class MediaTypeTest extends AbstractTypeTest
 
     public function testMissingFormContextOption(): void
     {
-        $this->mediaPool->expects($this->any())->method('getProviderList')->will($this->returnValue([
+        $this->mediaPool->expects($this->any())->method('getProviderList')->willReturn([
             'provider_a' => 'provider_a',
             'provider_b' => 'provider_b',
-        ]));
-        $this->mediaPool->expects($this->any())->method('getContexts')->will($this->returnValue([
+        ]);
+        $this->mediaPool->expects($this->any())->method('getContexts')->willReturn([
             'video' => [],
             'pic' => [],
-        ]));
+        ]);
 
         $this->expectException(MissingOptionsException::class);
 
@@ -84,14 +84,14 @@ class MediaTypeTest extends AbstractTypeTest
 
     public function testMissingFormProviderOption(): void
     {
-        $this->mediaPool->expects($this->any())->method('getProviderList')->will($this->returnValue([
+        $this->mediaPool->expects($this->any())->method('getProviderList')->willReturn([
             'provider_a' => 'provider_a',
             'provider_b' => 'provider_b',
-        ]));
-        $this->mediaPool->expects($this->any())->method('getContexts')->will($this->returnValue([
+        ]);
+        $this->mediaPool->expects($this->any())->method('getContexts')->willReturn([
             'video' => [],
             'pic' => [],
-        ]));
+        ]);
 
         $this->expectException(MissingOptionsException::class);
 
@@ -102,14 +102,14 @@ class MediaTypeTest extends AbstractTypeTest
 
     public function testInvalidFormProviderOption(): void
     {
-        $this->mediaPool->expects($this->any())->method('getProviderList')->will($this->returnValue([
+        $this->mediaPool->expects($this->any())->method('getProviderList')->willReturn([
             'provider_a' => 'provider_a',
             'provider_b' => 'provider_b',
-        ]));
-        $this->mediaPool->expects($this->any())->method('getContexts')->will($this->returnValue([
+        ]);
+        $this->mediaPool->expects($this->any())->method('getContexts')->willReturn([
             'video' => [],
             'pic' => [],
-        ]));
+        ]);
 
         $this->expectException(InvalidOptionsException::class);
         $this->expectExceptionMessage(
@@ -124,14 +124,14 @@ class MediaTypeTest extends AbstractTypeTest
 
     public function testInvalidFormContextOption(): void
     {
-        $this->mediaPool->expects($this->any())->method('getProviderList')->will($this->returnValue([
+        $this->mediaPool->expects($this->any())->method('getProviderList')->willReturn([
             'provider_a' => 'provider_a',
             'provider_b' => 'provider_b',
-        ]));
-        $this->mediaPool->expects($this->any())->method('getContexts')->will($this->returnValue([
+        ]);
+        $this->mediaPool->expects($this->any())->method('getContexts')->willReturn([
             'video' => [],
             'pic' => [],
-        ]));
+        ]);
 
         $this->expectException(InvalidOptionsException::class);
         $this->expectExceptionMessage(