+++ tests/Provider/AbstractProviderTest.php
@@ -50,21 +50,21 @@ abstract class AbstractProviderTest extends TestCase
         $this->formMapper
             ->expects($this->any())
             ->method('add')
-            ->will($this->returnCallback(function ($name, $type = null): void {
+            ->willReturnCallback(function ($name, $type = null): void {
                 if (null !== $type) {
                     $this->assertTrue(class_exists($type), sprintf('Unable to ensure %s is a FQCN', $type));
                 }
-            }));
+            });
 
         $this->formBuilder = $this->createMock(FormBuilder::class);
         $this->formBuilder
             ->expects($this->any())
             ->method('add')
-            ->will($this->returnCallback(function ($name, $type = null): void {
+            ->willReturnCallback(function ($name, $type = null): void {
                 if (null !== $type) {
                     $this->assertTrue(class_exists($type), sprintf('Unable to ensure %s is a FQCN', $type));
                 }
-            }));
+            });
 
         $this->formBuilder->expects($this->any())->method('getOption')->willReturn('api');